requirements.txt
tensorflow-cpu
numpy
pillow
streamlit